+++ app/models/concerns/valuable.rb
@@ -1,18 +1,16 @@
 # frozen_string_literal: true
 
-require 'valuables'
-
 module Valuable
   extend ActiveSupport::Concern
 
-  include Valuables, DateAndTimeParser
+  include DateAndTimeParser
 
   included do
     # Named Scopes
-    scope :with_variable_type, lambda { |arg| where( "#{self.table_name}.variable_id in (SELECT variables.id from variables where variables.variable_type IN (?))", arg ) }
+    scope :with_variable_type, lambda { |arg| where("#{table_name}.variable_id in (SELECT variables.id from variables where variables.variable_type IN (?))", arg) }
 
     # Model Validation
-    validates_presence_of :variable_id
+    validates :variable_id, presence: true
 
     # Model Relationships
     belongs_to :variable
@@ -25,23 +23,22 @@ module Valuable
     class_foreign_key = "#{self.class.name.underscore}_id".to_sym
 
     # Could use pluck, but pluck has issues with scopes and unsaved objects
-    old_response_ids = self.responses.collect { |r| r.id }
+    old_response_ids = responses.collect(&:id)
 
-    new_responses = []
-    values.select{|v| not v.blank?}.each do |value|
-      new_responses << Response.where(class_foreign_key => self.id, sheet_id: sheet.id, variable_id: self.variable_id, value: value).first_or_create( user_id: (current_user ? current_user.id : nil) )
+    self.responses = values.select(&:present?).collect do |value|
+      Response.where(class_foreign_key => id,
+                     sheet_id: sheet.id,
+                     variable_id: variable_id,
+                     value: value)
+              .first_or_create(user_id: (current_user ? current_user.id : nil))
     end
-    self.responses = new_responses
     Response.where(id: old_response_ids, class_foreign_key => nil).destroy_all
-  end
-
-  def get_response(raw_format = :raw)
-    Valuables.for(self).send(raw_format)
   end
 
   # Returns response as a hash that can sent to update_attributes
   def format_response(variable_type, response)
-    case variable_type when 'file'
+    case variable_type
+    when 'file'
       response = {} if response.blank?
     when 'date'
       month = parse_integer(response[:month])