+++ config/initializers/app_config.rb
@@ -7,4 +7,3 @@ require 'zip'
 require 'uri'
 # In lib directory
 require 'date_and_time_parser'
-require 'valuables'